backend/requirements.txt
# Core FastAPI stack
fastapi
uvicorn[standard]
pydantic-settings
python-multipart

# Database integration
appwrite

# Research AI Services (Phase 5)
exa-py
cerebras-cloud-sdk

# Voice Agent Services (Phase 3-4)
livekit-agents
livekit-agents[deepgram,silero,openai]
livekit-api

# Development and testing
pytest
pytest-asyncio
